Move determination of local projects into runbits package.
