+++ internal/runners/use/use.go
@@ -17,6 +17,7 @@ import (
 	"github.com/ActiveState/cli/internal/primer"
 	"github.com/ActiveState/cli/internal/prompt"
 	"github.com/ActiveState/cli/internal/runbits"
+	runbitsProject "github.com/ActiveState/cli/internal/runbits/project"
 	"github.com/ActiveState/cli/internal/subshell"
 	"github.com/ActiveState/cli/pkg/cmdlets/checker"
 	"github.com/ActiveState/cli/pkg/cmdlets/checkout"
@@ -78,9 +79,9 @@ func (u *Use) Run(params *Params) error {
 
 	checker.RunUpdateNotifier(u.svcModel, u.out)
 
-	proj, err := project.FromNamespaceLocal(params.Namespace, u.config, u.prompt)
+	proj, err := runbitsProject.FromNamespaceLocal(params.Namespace, u.config, u.prompt)
 	if err != nil {
-		if !project.IsLocalProjectDoesNotExistError(err) {
+		if !runbitsProject.IsLocalProjectDoesNotExistError(err) {
 			return locale.WrapError(err, "err_use", "Unable to use project")
 		}
 		if params.Namespace.Owner == "" {
